refactor(math): split _stream.py and lift deferred imports to module level
Addresses the two quality findings from the v1.1.3 boost scorecard.

#578 (code complexity): split the 1061-LOC _stream.py into a focused
façade (494 LOC) plus cohesive sibling modules, each well under 600 LOC:
  - _stream_state.py  — _StreamState (+ persist), StepResult, format version
  - _stream_math.py   — pure EWMA recurrences + input validation
  - _stream_solve.py  — per-step EWM / sliding-window position solvers
  - _stream_io.py     — save/load archive (BasanosStream.save/load delegate)

#579 (architecture): remove the two function-local runtime imports that
existed only to dodge import cycles.
  - _stream.py now imports BasanosEngine at module level (no real cycle).
  - _config_report.py reads type(config).model_fields instead of importing
    BasanosConfig at runtime, so _config.py can import ConfigReport at module
    level. The math package now has zero deferred runtime imports.

Behaviour unchanged: all 708 tests pass at 100% coverage; ty + mypy --strict,
docs-coverage (100%), deptry, security, and validate all green. radon shows no
C-or-worse blocks (avg A). Test patch/import targets updated to the new module
locations.
